pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "net-audit-automation"
version = "0.1.0"
description = "Authorized internal network audit automation toolkit"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0"
]
